requirements_dev.txt
black==24.3.0
flake8==6.0.0
importlib-metadata==6.0.0
mypy==0.991
pre-commit==2.20.0
pylint==2.15.9
pytest-cov==4.0.0
pytest-watch==4.2.0
ruff==0.0.217
sphinx-rtd-theme==1.2.0
sphinxcontrib-websupport==1.2.4
tox==4.0.19
types-mock==4.0.15.2
types-toml==0.10.8.5
wheel==0.38.4
